chore: cleanup shard_dimension_tracking.go
Removed code assosicated with shard-level metrics _publishing_;
shards should only respond to polling and don't need to push any
metrics themselves, except on shutdown/drop. Updated test cases
accordingly.

Moved some test helpers to helper_for_test.go, as the original
location has go:build integrationTest build tag.
